requirements.txt
Django==1.11.5
django-taggit==0.22.1
django-tinymce==2.6.0
pkg-resources==0.0.0
psycopg2==2.7.3.1
pytz==2017.2